+++ app/components/Checkbox/index.js
@@ -21,12 +21,31 @@ function Checkbox(props) {
       marginLeft: '0.5em',
       verticalAlign: 'middle',
     },
+    checkbox: {
+      height: '14px',
+      width: '14px',
+    },
+  };
+
+  const handleChange = (event) => {
+    if (props.cb) {
+      props.cb(event.target.checked, event);
+    }
   };
 
   return (
     <span style={props.style}>
-      <input id={props.id} type="checkbox" checked={props.checked} onChange={(e) => props.cb(e.target.checked)} />
-      <span style={styles.base}> {formatMessage(props.text)} </span>
+      <input
+        style={styles.checkbox}
+        id={props.id}
+        name={props.name}
+        type="checkbox"
+        checked={props.checked}
+        disabled={!props.cb}
+        onBlur={props.onBlur}
+        onChange={handleChange}
+      />
+      {props.text ? <span style={styles.base}> {formatMessage(props.text)} </span> : null}
     </span>
   );
 }
@@ -34,11 +53,13 @@ function Checkbox(props) {
 
 Checkbox.propTypes = {
   intl: intlShape.isRequired,
-  text: PropTypes.object.isRequired,
-  cb: PropTypes.func.isRequired,
+  name: PropTypes.string,
+  text: PropTypes.object,
+  cb: PropTypes.func,
   checked: PropTypes.bool.isRequired,
   style: PropTypes.object,
   id: PropTypes.string.isRequired,
+  onBlur: PropTypes.func,
 };
 
 export default injectIntl(Radium(Checkbox));